tracing: change begin(name, address) to use static storage duration
To optimize code hot paths for tracing we are trying to remove dynamic
allocations and memory copying as much as possible. One of the biggest
contributor to the performance overhead in tracing is string formatting.

With this patch we store a <const char *> instead of a copy of the
session name to reduce unwanted allocations and copies during request
processing.

Signed-off-by: Ivan Prisyazhnyy <ivan@scylladb.com>