tracing: ctx params inline vectors
To optimize code hot paths for tracing we are trying to remove dynamic
allocations and memory copying as much as possible. One of the biggest
contributor to the performance overhead in tracing is string formatting.

Not all tracing sessions are gonna to be written but all of them
fulfilled with the context parameters. There is a set of parameters that
is always filled. In most of the code hot paths there will be a query
and parameters to save. Most of the queries we expect to be non batch,
have 1 query and at most 2 parameters. That's why we can save a few
allocations here by changing the usage of std::vector to small_vector's.

This patch introduces use of small_vectors to minimize num of allocs in
code hot paths. small_vectors use local storage for storing objects
until their capacity is enough.

Signed-off-by: Ivan Prisyazhnyy <ivan@scylladb.com>