db: add checking for sufficient live nodes for iterators
The helper function counting if there are sufficient live nodes
used to operate only on full vectors, but the whole infrastructure
is already based on ranges, so it's easy to provide a specialized
variant of it that works on iterators. This variant will be useful
in later patches to check for sufficient live nodes for a slice
of a vector.