logalloc: bound the latency of reclaim using a proportional reserve
This patch trades a fixed percentage of memory for an upper bound on the
worst case latency of reclaim.

Today, that latency is unbounded. In the worst case, when the utilization of LSA
is full and the order of eviction is random, up to 15% of all memory may need to
be evicted before compactible segments (segments with <85% utilization) appear.

We can force a bound by making sure that the amount of memory released as clear
segments is proportional to the amount of memory evicted. To do that, we force
a reserve of clear segments proportional in size to the amount of currently used
(i.e. not holes) memory.

We pick a k and preserve the following invariant: there is a clear segment in
the reserve for every (k*segment_size) of used LSA memory.

In other words, we pay a fixed-fraction memory tax (1/k of used memory), but it
buys us a guarantee that freeing k*segment_size of memory by eviction will
release a clear segment, no matter how that freed memory is scattered, and no
matter how well compaction is working.

Implementation details:
We keep a shard-global counter. On alloc(n), we add n to this counter, on
free(n), we subtract n. When the counter rises above (k*segment_size), we add a
segment to the reserve and subtract (k*segment_size). When the counter falls
below 0, we release a segment from the reserve and add (k*segment_size).

This way, we maintain that:
(used memory) == (counter value) + (number of reserve segments) * (k * segment_size),
and 0 <= (counter value) <= (k * segment_size).

The patch includes a slight modification: it add some hysteresis to the counter
to avoid a situation where a series of small alloc/free repetitively crosses the
threshold and causes multiple segment allocations/frees. (Though since the LSA
doesn't immediately release segments and only marks them as free, repeatedly
freeing and allocating a segment is cheap and this optimization is mostly
redundant). To do that, we add segments to the reserve when the counter passes
not (k*segment_size), but (k*segment_size + antithrashing_margin).

This still ensures that:
(used memory) == (counter value) + (number of reserve segments) * (k * segment_size),
and 0 <= (counter value) <= (k * segment_size + antithrashing_margin),
but it also ensures that a reserve segment is allocated/freed at most once per
antithrashing_margin worth of allocs/frees.