repair: create repair_info before waiting for hints
User-requested repair should fail if a table that does not exist is
requested to be repaired. However, if the table which is being repaired
is dropped while repair is in progress, repair should handle that
gracefully and not fail.

Up to the point when `repair_info` objects are created, repair can fail
due to a nonexistent table. This point is usually reached very quickly.
However, in #8452 a feature was merged which waits for hints to be
replayed towards repaired nodes, and this wait delays the point at which
`repair_info`s are created. For example, if replaying hints lasts 60
seconds but the table is deleted 30 seconds after `nodetool repair` was
started, then the repair operation will fail.

In order to prevent such situations and handle the deleted table
gracefully, the code which can fail due to a nonexistent table is moved
before waiting for hints to be replayed.

Fixes #8612

Tests:
    unit(dev, debug)
    dtest(repair_additional_test.py:RepairAdditionalTest.\
        repair_while_table_is_dropped_test)